+++ cnxarchive/tests/views/test_extras.py
@@ -543,3 +543,18 @@ application problems.</div>""",
                            u'url': u'http://creativecommons.org/licenses/by-nc-sa/4.0/',
                            u'version': u'4.0'}],
             })
+        expected_messages = [
+            {u'message': u'This site is scheduled to be down for maintaince, please excuse the interuption. Thank you.',
+             u'name': u'Maintenance',
+             u'priority': 1},
+            {u'message': u"We have free books at free prices! Don't miss out!",
+             u'name': u'Notice',
+             u'priority': 8}
+        ]
+
+        def _remove_timestamps(messages):
+            for message in messages:
+                message.pop('starts')
+                message.pop('ends')
+            return messages
+        self.assertEqual(expected_messages, _remove_timestamps(messages))
