+++ cnxarchive/tests/views/test_search_views.py
@@ -37,15 +37,6 @@ SEARCH_RESULTS_FILEPATH = os.path.join(testing.DATA_DIRECTORY,
 with open(SEARCH_RESULTS_FILEPATH, 'r') as file:
     SEARCH_RESULTS = json.load(file)
 
-expected_messages = [
-    {u'message': u'This site is scheduled to be down for maintaince, please excuse the interuption. Thank you.',
-     u'name': u'Maintenance',
-     u'priority': 1},
-    {u'message': u"We have free books at free prices! Don't miss out!",
-     u'name': u'Notice',
-     u'priority': 8}
-]
-
 
 @mock.patch('cnxarchive.views.search_view.fromtimestamp', mock.Mock(side_effect=testing.mocked_fromtimestamp))
 class ViewsTestCase(unittest.TestCase):
@@ -927,10 +918,3 @@ class ViewsTestCase(unittest.TestCase):
 
         self.assertEqual(results['results']['total'], 7)
         self.assertEqual(self.db_search_call_count, 2)
-
-        def _remove_timestamps(messages):
-            for message in messages:
-                message.pop('starts')
-                message.pop('ends')
-            return messages
-        self.assertEqual(expected_messages, _remove_timestamps(messages))
